add optional callback for decrement and resetKey methods
