+++ lib/mongoStore.js
@@ -3,6 +3,7 @@
 var _ = require('underscore');
 var Steppy = require('twostep').Steppy;
 var MongoClient = require('mongodb').MongoClient;
+var utils = require('./utils');
 
 var MongoStore = function(options) {
 	var allOptions = _({}).defaults(
@@ -160,6 +161,10 @@ MongoStore.prototype.incr = function(key, callback) {
 
 MongoStore.prototype.decrement = function(key) {
 	var self = this;
+	var callback = utils.getCallbackFromArgs(
+		_(arguments).toArray()
+	);
+
 	Steppy(
 		function() {
 			self._getCollection(this.slot());
@@ -192,12 +197,20 @@ MongoStore.prototype.decrement = function(key) {
 			if (err) {
 				self.errorHandler(err);
 			}
+
+			if (callback) {
+				callback(err);
+			}
 		}
 	);
 };
 
 MongoStore.prototype.resetKey = function(key) {
 	var self = this;
+	var callback = utils.getCallbackFromArgs(
+		_(arguments).toArray()
+	);
+
 	Steppy(
 		function() {
 			self._getCollection(this.slot());
@@ -211,6 +224,10 @@ MongoStore.prototype.resetKey = function(key) {
 		function(err) {
 			if (err) {
 				self.errorHandler(err);
+			}
+
+			if (callback) {
+				callback(err);
 			}
 		}
 	);
