+++ lib/utils/getCallbackFromArgs.js
@@ -1,0 +1,15 @@
+'use strict';
+
+var _ = require('underscore');
+
+module.exports = function(args) {
+	var callback;
+	var lastArg = _(args).last();
+	var callback;
+
+	if (args.length > 1 && _(lastArg).isFunction()) {
+		callback = lastArg;
+	}
+
+	return callback;
+};
