+++ lib/utils/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+exports.getCallbackFromArgs = require('./getCallbackFromArgs');
