+++ test/mongoStore/_createCollection/withMongoClient2.x.x.js
@@ -6,7 +6,7 @@ var rewire = require('rewire');
 var _ = require('underscore');
 var testUtils = require('./utils');
 
-var MongoStore = rewire('../../lib/mongoStore');
+var MongoStore = rewire('../../../lib/mongoStore');
 
 var describeTitle = 'MongoClient.prototype._createCollection ' +
 	'with MongoClient version 2.x.x';
