+++ test/mongoStore/_getCollection/withNotPreparedCollectionStateAndCollection.js
@@ -6,19 +6,19 @@ var rewire = require('rewire');
 var _ = require('underscore');
 var testUtils = require('./utils');
 
-var MongoStore = rewire('../../lib/mongoStore');
+var MongoStore = rewire('../../../lib/mongoStore');
 
 var describeTitle = 'MongoStore.prototype._getCollection ' +
-	'with collection.createIndex error';
+	'with "notPrepared" collectionState and collection';
 describe(describeTitle, function() {
 	var testData = {
 		mongoStoreContext: {
 			_collectionState: 'notPrepared'
-		},
-		collectionCreateIndexError: new Error('test createIndex error')
+		}
 	};
 
-	var mocks = testUtils.getMocks(testData);
+	var mocks = testUtils.getMocks();
+	mocks._dynamic.mongoStoreContext.collection = mocks._dynamic.collection;
 
 	var revertMocks;
 
@@ -28,7 +28,7 @@ describe(describeTitle, function() {
 		);
 	});
 
-	it('should throw error', function(done) {
+	it('should return collection', function(done) {
 		Steppy(
 			function() {
 				MongoStore.prototype._getCollection.call(
@@ -39,20 +39,15 @@ describe(describeTitle, function() {
 					this.slot()
 				);
 			},
-			function(err) {
-				expect(err).eql(
-					testData.collectionCreateIndexError
+			function(err, collection) {
+				expect(collection).eql(
+					mocks._dynamic.mongoStoreContext.collection
 				);
 
-				done();
-			}
+				this.pass(null);
+			},
+			done
 		);
-	});
-
-	it('mongoStore._collectionState should be "notPrepared"', function() {
-		expect(
-			testData.mongoStoreContext._collectionState
-		).eql('notPrepared');
 	});
 
 	it('setImmediate should not be called', function() {
@@ -65,22 +60,16 @@ describe(describeTitle, function() {
 		).eql(0);
 	});
 
-	it('_createCollection should be called', function() {
+	it('_createCollection should not be called', function() {
 		expect(
 			mocks._dynamic.mongoStoreContext._createCollection.callCount
-		).eql(1);
-
-		var createCollectionArgs = mocks._dynamic.mongoStoreContext
-			._createCollection.args[0];
-
-		expect(createCollectionArgs).length(1);
-		expect(createCollectionArgs[0]).a('function');
+		).eql(0);
 	});
 
 	it(
 		'collection.createIndex should be called for setting ttl index',
 		function() {
-			var collectionMock = mocks._dynamic.collection;
+			var collectionMock = mocks._dynamic.mongoStoreContext.collection;
 
 			expect(collectionMock.createIndex.callCount).eql(1);
