+++ test/mongoStore/incr/withFindOneAndUpdateError.js
@@ -6,7 +6,7 @@ var rewire = require('rewire');
 var _ = require('underscore');
 var testUtils = require('./utils');
 
-var MongoStore = rewire('../../lib/mongoStore');
+var MongoStore = rewire('../../../lib/mongoStore');
 
 var describeTitle = 'MongoStore.prototype.incr with findOneAndUpdate error';
 describe(describeTitle, function() {
