+++ test/mongoStore/incr/withResetExpireDateOnCange.js
@@ -5,7 +5,7 @@ var rewire = require('rewire');
 var _ = require('underscore');
 var testUtils = require('./utils');
 
-var MongoStore = rewire('../../lib/mongoStore');
+var MongoStore = rewire('../../../lib/mongoStore');
 
 var describeTitle = 'MongoStore.prototype.incr with resetExpireDateOnChange';
 describe(describeTitle, function() {
