+++ test/mongoStore/resetKey/withSuitableParams.js
@@ -1,11 +1,12 @@
 'use strict';
 
 var expect = require('expect.js');
+var Steppy = require('twostep').Steppy;
 var rewire = require('rewire');
 var _ = require('underscore');
 var testUtils = require('./utils');
 
-var MongoStore = rewire('../../lib/mongoStore');
+var MongoStore = rewire('../../../lib/mongoStore');
 
 var describeTitle = 'MongoStore.prototype.resetKey with suitable params';
 describe(describeTitle, function() {
@@ -14,17 +15,19 @@ describe(describeTitle, function() {
 	var mocks = testUtils.getMocks(testData);
 
 	it('should be ok', function(done) {
-		MongoStore.prototype.resetKey.call(
-			_({}).extend(
-				testData.mongoStoreContext,
-				mocks._dynamic.mongoStoreContext
-			),
-			testData.key
+		Steppy(
+			function() {
+				MongoStore.prototype.resetKey.call(
+					_({}).extend(
+						testData.mongoStoreContext,
+						mocks._dynamic.mongoStoreContext
+					),
+					testData.key,
+					this.slot()
+				);
+			},
+			done
 		);
-
-		setTimeout(function() {
-			done();
-		}, 10);
 	});
 
 	it('_getCollection should be called', function() {
